Fix sum logic when adding cards
The sum of the cards were adding on top of the sum of the previous
cards. I added sum = 0 in the renderGame function so every time the
function is called, sum resets to 0. This resets the sum and only adds
values in the array after each press of a button